move database code from utils
