+++ src/test/database.ts
@@ -3,7 +3,7 @@ import config from 'config';
 import { expect } from 'chai';
 import Datastore from 'nedb';
 
-import { Database } from '../backend/utils/database/index.js';
+import { Database } from '../backend/database/index.js';
 
 interface Document {
   data?: any;
